--- a/src/clojure_lsp/crawler.clj
+++ b/src/clojure_lsp/crawler.clj
@@ -203,7 +203,7 @@
                       [uri (safe-find-references uri text false true)])))
              (remove (comp nil? second)))
         output-chan (async/chan 5)]
-    (async/pipeline-blocking 5 output-chan xf (async/to-chan jars) true (fn [e] (log/warn e "hello")))
+    (async/pipeline-blocking 5 output-chan xf (async/to-chan! jars) true (fn [e] (log/warn e "hello")))
     (async/<!! (async/into {} output-chan))))
 
 (defn crawl-source-dirs [dirs]
@@ -219,7 +219,7 @@
                                   (safe-find-references uri (slurp uri) false false))))
             (remove (comp nil? second)))
         output-chan (async/chan)]
-    (async/pipeline-blocking 5 output-chan xf (async/to-chan dirs) true (fn [e] (log/warn e "hello")))
+    (async/pipeline-blocking 5 output-chan xf (async/to-chan! dirs) true (fn [e] (log/warn e "hello")))
     (async/<!! (async/into {} output-chan))))
 
 (defn lookup-classpath [root-path command-args env]