#fix 208

This PR changes how we save usages on db. We now save `file-type` as a set, so instead of 2 usages with `clj` and another with `cljs` file-types, we now persist only one with `#{clj cljs}` as file-type.
This is only applicable for non `norename` tags as we just want to do that for user namespace/alias usages