@bpringe 2020-12-24
I'm not sure what in these changes might cause this, but I noticed when the lsp db doesn't exist in a project, it's taking 90+ seconds to return the initialize response. Before these changes, it took only about 40k seconds. I tested at least a couple times with a jar from before these changes and a jar with these changes.

Even with the changes though, if the db exists already, it only takes about 6 seconds.

Any idea how these changes might have more than doubled the initialize response time when no lsp db exists?