@ericdallo 2020-12-24
Hum, I don't think this PR could cause that, we just iterate a single time to each usage when persisting which seems to not be a problem.
The initial delay is already a problem since we started to use clj-kondo.
I recommend testing it checking the initialize time displayed on `/tmp/lsp.out` and removing `/your-project/.lsp` folder to reanalyze without any cache