@ericdallo 2020-12-24
You can also wrap it with a  `log/debug` with the time spent in that function for debugging purposes