@bpringe 2020-12-24
> I recommend testing it checking the initialize time displayed on /tmp/lsp.out and removing /your-project/.lsp folder to

I did this and used the trace logging which says how long each request/response takes. Before these changes, with the .lsp folder removed it took ~30 seconds to respond, and with these changes with the .lsp folder removed, it takes about ~90 seconds to respond.

I agree these changes don't seem like they should cause an issue, just reporting what I've seen. Not sure if anything should be done about it.