@ericdallo 2021-01-09
Hey, after a lot of tests I could prove that this PR indeed introduced a regression regarding performance :(
But I could improve it to make it almost as fast as before:

Before the cljc duplicated usages fix:
```clojure
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 847.996168 ms
    Execution time std-deviation : 42.774901 ms
   Execution time lower quantile : 801.961730 ms ( 2.5%)
   Execution time upper quantile : 913.601543 ms (97.5%)
                   Overhead used : 1.947168 ns
```

[code before](https://github.com/clojure-lsp/clojure-lsp/blob/master/src/clojure_lsp/parser.clj#L1131-L1157) the performance fix.
```clojure
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 3.575516 sec
    Execution time std-deviation : 224.909583 ms
   Execution time lower quantile : 3.304419 sec ( 2.5%)
   Execution time upper quantile : 3.841209 sec (97.5%)
                   Overhead used : 1.947168 ns
```

[With the performance fix](https://github.com/clojure-lsp/clojure-lsp/pull/243)
```clojure
             Execution time mean : 877.753781 ms
    Execution time std-deviation : 12.686530 ms
   Execution time lower quantile : 857.340680 ms ( 2.5%)
   Execution time upper quantile : 890.014269 ms (97.5%)
                   Overhead used : 1.947168 ns
```

This function is used every time we need to crawl a code in a file to parse and understand the symbols, since we could have a lot of file for multiple jars/source-dirs, this should improve the performance as it was before :)

note: I did the performance tests using [criterium](https://github.com/hugoduncan/criterium) benchmark tool.